% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_visualize_ensemble_step.R
\name{custom_visualize_ensemble}
\alias{custom_visualize_ensemble}
\title{Custom Ensemble Visualization Step}
\arguments{
\item{seurat_object}{A Seurat object containing the data to visualize.}
}
\value{
A combined ggplot object representing the ensemble visualization.
}
\description{
This function generates customized UMAP visualizations and combines them into a grid layout.
It includes options for coloring by ensemble output, ScType classifications, and other metadata.
}
